Rename Annotation namespace to Attribute
